// frontend/src/types/context/auth.ts
export interface IRegistration {
  full_name: string | undefined;
  first_name: string | undefined;
  last_name?: string | undefined;
  user_name?: string;
  email?: string;
  phone?: string;
  referral_code?: string;
  // passport_username?: string; 
  password: string;
  user_type: "Agent" | "Student" | "Support";
  registration_source: "email" | "phone" | "facebook" | "google";
  login_source: "email" | "phone" | "facebook" | "google";
}
// ...
export interface ILoginRequest {
  password: string | undefined;
  email?: string | undefined;
  phone?: string | undefined;
  remember_me?: boolean;
}